fix infinite loop error message sometimes always coming up on startup

+++ src/gui/interface/Engine.h
@@ -35,8 +35,8 @@ namespace ui
 		void Begin(int width, int height);
 		inline bool Running() { return running_; }
 		inline bool Broken() { return break_; } 
-		inline int LastTick() { return lastTick; }
-		inline void LastTick(int tick) { lastTick = tick; }
+		inline long unsigned int LastTick() { return lastTick; }
+		inline void LastTick(long unsigned int tick) { lastTick = tick; }
 		void Exit();
 		void Break();
 		void UnBreak();
@@ -92,7 +92,7 @@ namespace ui
 		bool break_;
 		bool FastQuit;
 		
-		int lastTick;
+		long unsigned int lastTick;
 		int mouseb_;
 		int mousex_;
 		int mousey_;
